--- flow-typedefs/MessageText.js.flow
@@ -1,23 +1,0 @@
-// @flow
-import { Text } from 'react-native'
-import type { LeftRightStyle, IMessage } from './types'
-import type {
-  TextStyleProp,
-  ViewStyleProp,
-} from 'react-native/Libraries/StyleSheet/StyleSheet'
-import * as React from 'react'
-
-export type MessageTextProps<TMessage: IMessage = IMessage> = $ReadOnly<{|
-  position: 'left' | 'right',
-  currentMessage: TMessage,
-  containerStyle: LeftRightStyle<ViewStyleProp>,
-  textStyle?: LeftRightStyle<TextStyleProp>,
-  linkStyle?: LeftRightStyle<TextStyleProp>,
-  textProps?: $Shape<React.ElementProps<typeof Text>>,
-  customTextStyle?: TextStyleProp,
-  parsePatterns?: TextStyleProp => any,
-|}>
-
-export default class MessageText<
-  TMessage: IMessage = IMessage,
-> extends React.Component<MessageTextProps<TMessage>> {}
