--- flow-typedefs/QuickReplies.js.flow
@@ -1,25 +1,0 @@
-// @flow
-import * as React from 'react'
-import type { IMessage, Reply } from './types'
-import type { ViewStyleProp, TextStyleProp } from 'react-native/Libraries/StyleSheet/StyleSheet'
-
-export type QuickRepliesProps<TMessage: IMessage = IMessage> = $ReadOnly<{|
-  nextMessage?: TMessage,
-  currentMessage: TMessage,
-  color?: string,
-  sendText?: string,
-  quickReplyStyle?: ViewStyleProp,
-  quickReplyTextStyle?: TextStyleProp,
-  quickReplyContainerStyle?: ViewStyleProp,
-  onQuickReply: (Array<Reply>) => void,
-  renderQuickReplySend?: () => React.Node,
-|}>
-
-export type QuickRepliesState = {|
-  replies: Array<Reply>,
-|}
-
-export default class QuickReplies extends React.Component<
-  QuickRepliesProps,
-  QuickRepliesState,
-> {}
