--- flow-typedefs/index.js.flow
@@ -1,2 +1,0 @@
-// @flow
-export * from './GiftedChat'
