+++ src/Bubble/styles.ts
@@ -1,0 +1,73 @@
+import { StyleSheet } from 'react-native'
+import Color from '../Color'
+
+const styles = {
+  left: StyleSheet.create({
+    container: {
+      alignItems: 'flex-start',
+    },
+    wrapper: {
+      borderRadius: 15,
+      backgroundColor: Color.leftBubbleBackground,
+      marginRight: 60,
+      minHeight: 20,
+      justifyContent: 'flex-end',
+    },
+    containerToNext: {
+      borderBottomLeftRadius: 3,
+    },
+    containerToPrevious: {
+      borderTopLeftRadius: 3,
+    },
+    bottom: {
+      flexDirection: 'row',
+      justifyContent: 'flex-start',
+    },
+  }),
+  right: StyleSheet.create({
+    container: {
+      alignItems: 'flex-end',
+    },
+    wrapper: {
+      borderRadius: 15,
+      backgroundColor: Color.defaultBlue,
+      marginLeft: 60,
+      minHeight: 20,
+      justifyContent: 'flex-end',
+    },
+    containerToNext: {
+      borderBottomRightRadius: 3,
+    },
+    containerToPrevious: {
+      borderTopRightRadius: 3,
+    },
+    bottom: {
+      flexDirection: 'row',
+      justifyContent: 'flex-end',
+    },
+  }),
+  content: StyleSheet.create({
+    tick: {
+      fontSize: 10,
+      backgroundColor: Color.backgroundTransparent,
+      color: Color.white,
+    },
+    tickView: {
+      flexDirection: 'row',
+      marginRight: 10,
+    },
+    username: {
+      top: -3,
+      left: 0,
+      fontSize: 12,
+      backgroundColor: Color.backgroundTransparent,
+      color: '#aaa',
+    },
+    usernameView: {
+      flexDirection: 'row',
+      marginHorizontal: 10,
+    },
+  }),
+}
+
+export default styles
