+++ src/GiftedChat/styles.ts
@@ -1,0 +1,7 @@
+import { StyleSheet } from 'react-native'
+
+export default StyleSheet.create({
+  contentContainer: {
+    overflow: 'hidden',
+  },
+})
