+++ src/MessageContainer/components/Item/types.ts
@@ -1,5 +1,5 @@
 import { MessageContainerProps, DaysPositions } from '../../types'
-import { IMessage } from '../../../Models'
+import { IMessage } from '../../../types'
 
 export interface ItemProps extends MessageContainerProps<IMessage> {
   onRefDayWrapper: (ref: any, id: string | number, createdAt: number) => void
