+++ src/QuickReplies.tsx
@@ -1,5 +1,4 @@
 import React, { useState, useMemo, useCallback } from 'react'
-import PropTypes from 'prop-types'
 import {
   Text,
   StyleSheet,
@@ -9,9 +8,8 @@ import {
   ViewStyle,
   TextStyle,
 } from 'react-native'
-import { IMessage, Reply } from './Models'
+import { IMessage, Reply } from './types'
 import Color from './Color'
-import { StylePropType } from './utils'
 import { warning } from './logging'
 import stylesCommon from './styles'
 
@@ -89,6 +87,15 @@ export function QuickReplies ({
     return false
   }, [currentMessage, nextMessage])
 
+  const handleSend = useCallback((repliesData: Reply[]) => () => {
+    onQuickReply?.(
+      repliesData.map((reply: Reply) => ({
+        ...reply,
+        messageId: currentMessage!._id,
+      }))
+    )
+  }, [onQuickReply, currentMessage])
+
   const handlePress = useCallback(
     (reply: Reply) => () => {
       if (currentMessage) {
@@ -112,17 +119,8 @@ export function QuickReplies ({
         }
       }
     },
-    [replies, currentMessage]
+    [replies, currentMessage, handleSend]
   )
-
-  const handleSend = (repliesData: Reply[]) => () => {
-    onQuickReply?.(
-      repliesData.map((reply: Reply) => ({
-        ...reply,
-        messageId: currentMessage!._id,
-      }))
-    )
-  }
 
   if (!shouldComponentDisplay)
     return null
@@ -173,13 +171,4 @@ export function QuickReplies ({
       )}
     </View>
   )
-}
-
-QuickReplies.propTypes = {
-  currentMessage: PropTypes.object.isRequired,
-  onQuickReply: PropTypes.func,
-  color: PropTypes.string,
-  sendText: PropTypes.string,
-  renderQuickReplySend: PropTypes.func,
-  quickReplyStyle: StylePropType,
 }
